+++ src/main/java/opensource/onlinestore/model/entity/MessageEntity.java
@@ -1,5 +1,7 @@
 package opensource.onlinestore.model.entity;
 
+import opensource.onlinestore.model.enums.MessageType;
+
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
