+++ src/main/java/opensource/onlinestore/model/entity/UserEntity.java
@@ -1,5 +1,7 @@
 package opensource.onlinestore.model.entity;
 
+import opensource.onlinestore.model.enums.ActivityStatus;
+import opensource.onlinestore.model.enums.Role;
 import org.hibernate.validator.constraints.Email;
 import org.hibernate.validator.constraints.NotEmpty;
