+++ src/main/java/opensource/onlinestore/model/enums/ActivityStatus.java
@@ -1,4 +1,4 @@
-package opensource.onlinestore.model.entity;
+package opensource.onlinestore.model.enums;
 
 /**
  * Created by maks(avto12@i.ua) on 26.01.2016.
