+++ src/main/java/opensource/onlinestore/model/enums/MessageType.java
@@ -1,4 +1,4 @@
-package opensource.onlinestore.model.entity;
+package opensource.onlinestore.model.enums;
 
 /**
  * Created by malex on 31.01.16.
