+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 data.username=postgres
-data.password=1234
+data.password=2687484a
 data.jdbc.driver=org.postgresql.Driver
 data.jdbc.url=jdbc:postgresql://localhost:5432/store
 data.showsql=true
