+++ src/test/java/opensource/onlinestore/service/dbunit/RelationshipsUserMessageSupport.java
@@ -4,7 +4,7 @@ import com.github.springtestdbunit.DbUnitTestExecutionListener;
 import com.github.springtestdbunit.annotation.DatabaseSetup;
 import opensource.onlinestore.configuration.AppConfigTest;
 import opensource.onlinestore.model.entity.MessageEntity;
-import opensource.onlinestore.model.entity.MessageType;
+import opensource.onlinestore.model.enums.MessageType;
 import opensource.onlinestore.model.entity.UserEntity;
 import opensource.onlinestore.model.util.UserEntityUtil;
 import opensource.onlinestore.service.MessageService;
