+++ src/test/java/opensource/onlinestore/service/logic/MessageServiceImpLogicTest.java
@@ -3,7 +3,7 @@ package opensource.onlinestore.service.logic;
 import opensource.onlinestore.configuration.AppConfigTest;
 
 import opensource.onlinestore.model.entity.MessageEntity;
-import opensource.onlinestore.model.entity.MessageType;
+import opensource.onlinestore.model.enums.MessageType;
 import opensource.onlinestore.model.entity.UserEntity;
 import opensource.onlinestore.model.util.UserEntityUtil;
 import opensource.onlinestore.service.MessageService;
